refactor restaurand dynamic page
