+++ client/src/app/components/detailPage/mockdata.ts
@@ -1,21 +1,21 @@
-export interface RestaurantMockData {
-  name: string;
-  ratingNum: number;
-  reviews: number;
-  tags: string;
-  add: string;
-  map: string;
-}
+// export interface RestaurantMockData {
+//   name: string;
+//   ratingNum: number;
+//   reviews: number;
+//   tags: string;
+//   add: string;
+//   map: string;
+// }
 
-const restaurantMockData: RestaurantMockData[] = [
-  {
-    name: "Saku",
-    ratingNum: 5,
-    reviews: 120,
-    tags: "Japanese",
-    add: "567 Clarke Rd #107, Coquitlam, BC V3J 0K7",
-    map: "mockMap.png",
-  },
-];
+// const restaurantMockData: RestaurantMockData[] = [
+//   {
+//     name: "Saku",
+//     ratingNum: 5,
+//     reviews: 120,
+//     tags: "Japanese",
+//     add: "567 Clarke Rd #107, Coquitlam, BC V3J 0K7",
+//     map: "mockMap.png",
+//   },
+// ];
 
-export { restaurantMockData };
+// export { restaurantMockData };
