+++ client/src/app/detailpage/page.tsx
@@ -1,23 +1,23 @@
-"use client";
-import React, { useState } from "react";
-import RestaurantWithMap from "../components/detailPage/RestaurantWithMap";
-import PersonalReview from "../components/detailPage/PersonalReview";
-import NoReviews from "../components/detailPage/NoReviews";
+// "use client";
+// import React, { useState } from "react";
+// import RestaurantWithMap from "../components/detailPage/RestaurantWithMap";
+// import PersonalReview from "../components/detailPage/PersonalReview";
+// import NoReviews from "../components/detailPage/NoReviews";
 
-function Page() {
-  const [noReviews, setNoReviews] = useState(false);
-  // check if there's review, if yes, rendering PersonalReview, if not, rendering noReviews
+// function Page() {
+//   const [noReviews, setNoReviews] = useState(false);
+//   // check if there's review, if yes, rendering PersonalReview, if not, rendering noReviews
 
-  return (
-    <>
-      <div className="flex">
-        <div className="w-2/3">
-          {noReviews ? <NoReviews /> : <PersonalReview />}
-        </div>
-        <RestaurantWithMap noReviews={noReviews} />
-      </div>
-    </>
-  );
-}
+//   return (
+//     <>
+//       <div className="flex">
+//         <div className="w-2/3">
+//           {noReviews ? <NoReviews /> : <PersonalReview />}
+//         </div>
+//         <RestaurantWithMap noReviews={noReviews} />
+//       </div>
+//     </>
+//   );
+// }
 
-export default Page;
+// export default Page;
