+++ client/src/app/restaurants/[restaurantId]/page.tsx
@@ -7,15 +7,20 @@
 // export default page;
 
 "use client";
-import NoReviews from "@/app/components/detailPage/NoReviews";
-import PersonalReview from "@/app/components/detailPage/PersonalReview";
+
+import RestaurantWithMap from "@/app/components/detailPage/RestaurantWithMap";
 import Card from "@/app/components/restaurant/Card";
-import React, { useState } from "react";
-// import RestaurantWithMap from "../components/detailPage/RestaurantWithMap";
-// import PersonalReview from "../components/detailPage/PersonalReview";
-// import NoReviews from "../components/detailPage/NoReviews";
+import NoReviews from "@/app/components/review/NoReviews";
+import PersonalReview from "@/app/components/review/PersonalReview";
+import { DummyRestaurantData } from "@/components/common/types";
+// import { useParams, usePathname, useRouter } from "next/navigation";
+import { useParams } from "next/navigation";
+import { useState } from "react";
 
 function Page() {
+  const params = useParams();
+  // console.log("params", params);
+
   const [noReviews, setNoReviews] = useState(false);
   // check if there's review, if yes, rendering PersonalReview, if not, rendering noReviews
 
@@ -25,7 +30,10 @@ function Page() {
         <div className="w-2/3">
           {noReviews ? <NoReviews /> : <PersonalReview />}
         </div>
-        <Card />
+        <div>
+          <div className="text-red-800">restaurant {params.restaurantId}</div>
+          <RestaurantWithMap noReviews={noReviews} />
+        </div>
       </div>
     </>
   );
